Extract VSCode telemetry/recommendation utilities into common framework
- Move hot-event JSON parsing and aggregation helpers to ProfiledBuildScenario
- Move GHCP availability, recommendation parsing, and profile selection helpers to common base
- Move available-profile discovery and CPU/RTP/hot-event collection helpers to common base
- Reduce VSCodeScenario to scenario-specific hooks and behavior
- Update VSCode tests to patch shared framework module paths
