+++ macos/performance-profiles/demo_framework/scenarios/vscode.py
@@ -8,10 +8,8 @@ from pathlib import Path
 from typing import Optional, List
 import json
 import os
-import re
 import shutil
 import subprocess
-import threading
 import time
 
 from .base import ScenarioConfig
@@ -55,42 +53,6 @@ class VSCodeScenario(ProfiledBuildScenario):
         self.recommended_profiles = None
         self.recommendation_source = "default"
 
-    def _to_int(self, value):
-        """Best-effort int conversion for mixed numeric JSON fields."""
-        try:
-            return int(value)
-        except Exception:
-            try:
-                return int(float(value))
-            except Exception:
-                return 0
-
-    def _load_hot_event_entries(self, path: Path):
-        """Load and normalize hot-event-source entries from JSON artifact."""
-        if not path.exists() or path.stat().st_size == 0:
-            return []
-        try:
-            data = json.loads(path.read_text())
-            entries = data.get("eventSource", [])
-            normalized = []
-            for e in entries:
-                if not isinstance(e, dict):
-                    continue
-                auth = self._to_int(e.get("authCount", 0))
-                notify = self._to_int(e.get("notifyCount", 0))
-                normalized.append(
-                    {
-                        "path": e.get("path", "?"),
-                        "auth": auth,
-                        "notify": notify,
-                        "total": auth + notify,
-                    }
-                )
-            normalized.sort(key=lambda x: x["total"], reverse=True)
-            return normalized
-        except Exception:
-            return []
-
     def _hot_event_python_summary(self, before: Path, after: Optional[Path] = None):
         """Generate local Python summary for hot event sources."""
         before_entries = self._load_hot_event_entries(before)
@@ -117,31 +79,6 @@ class VSCodeScenario(ProfiledBuildScenario):
                 print(f"   Aggregate after:  auth={a_auth} notify={a_notify} total={a_auth + a_notify}")
                 delta = (a_auth + a_notify) - (b_auth + b_notify)
                 print(f"   Delta (after-before): {delta:+d} events")
-
-    def _hot_event_aggregate(self, path: Path):
-        """Return aggregate auth/notify/total counts from a hot-event artifact."""
-        entries = self._load_hot_event_entries(path)
-        if not entries:
-            return None
-        auth = sum(e["auth"] for e in entries)
-        notify = sum(e["notify"] for e in entries)
-        return {"auth": auth, "notify": notify, "total": auth + notify}
-
-    def _has_ghcp_cli(self):
-        """Return True when GitHub CLI Copilot command appears available."""
-        try:
-            if shutil.which("gh") is None:
-                return False
-            res = subprocess.run(
-                ["gh", "copilot", "--help"],
-                capture_output=True,
-                text=True,
-                timeout=5,
-                check=False,
-            )
-            return res.returncode == 0
-        except Exception:
-            return False
 
     def _hot_event_ghcp_analysis(self, before: Path, after: Optional[Path] = None):
         """Ask GitHub Copilot CLI for a textual interpretation of hot-event data."""
@@ -199,37 +136,6 @@ class VSCodeScenario(ProfiledBuildScenario):
         except Exception as e:
             print_info(f"GHCP analysis failed: {e}")
 
-    def _get_available_profiles(self) -> List[str]:
-        """Read available performance profiles from mdatp CLI with fallback."""
-        try:
-            result = subprocess.run(
-                ["sudo", "mdatp", "performance-profiles", "list-available"],
-                capture_output=True,
-                text=True,
-                timeout=10,
-                check=False,
-            )
-            lines = [line.strip() for line in (result.stdout or "").splitlines() if line.strip()]
-            parsed = []
-            for line in lines:
-                lower = line.lower()
-                if lower.startswith("merge policy:"):
-                    continue
-                if lower.startswith("no available performance profiles"):
-                    continue
-                if lower.startswith("available profiles"):
-                    continue
-                if line in ("---", "====================================="):
-                    continue
-
-                first = line.split()[0].strip("-*\t")
-                if first in self.config.profiles and first not in parsed:
-                    parsed.append(first)
-
-            return parsed or list(self.config.profiles)
-        except Exception:
-            return list(self.config.profiles)
-
     def _python_profile_recommendations(self, hot_events: Path, available_profiles: Optional[List[str]] = None) -> List[str]:
         """Derive profile recommendations from hot-event path heuristics."""
         entries = self._load_hot_event_entries(hot_events)
@@ -258,139 +164,6 @@ class VSCodeScenario(ProfiledBuildScenario):
         ]
         return ranked
 
-    def _ghcp_profile_recommendations(
-        self, hot_events: Path, available_profiles: Optional[List[str]] = None
-    ) -> List[str]:
-        """Ask GH Copilot CLI to pick profile names from known profile list."""
-        if not self._has_ghcp_cli():
-            return []
-
-        entries = self._load_hot_event_entries(hot_events)
-        if not entries:
-            return []
-
-        top = entries[:10]
-        top_text = "\n".join(
-            [f"- total={e['total']} auth={e['auth']} notify={e['notify']} path={e['path']}" for e in top]
-        )
-        allowed_profiles = list(available_profiles or self.config.profiles)
-        allowed = ", ".join(allowed_profiles)
-        prompt = (
-            "Analyze this MDE hot-event telemetry from a VS Code build and recommend which available "
-            "performance profiles should be applied.\n"
-            f"Available profiles: {allowed}.\n"
-            "Return your final recommendation in this exact machine-readable line:\n"
-            "RECOMMENDED_PROFILES: <comma-separated profile names from the available list>\n"
-            "You may include brief reasoning before that final line.\n\n"
-            f"Hot event sources:\n{top_text}\n"
-        )
-
-        try:
-            res = subprocess.run(
-                ["gh", "copilot", "suggest", "-t", "shell", prompt],
-                capture_output=True,
-                text=True,
-                timeout=60,
-                check=False,
-            )
-            if res.returncode != 0:
-                err_line = ""
-                for line in (res.stderr or "").splitlines():
-                    if line.strip():
-                        err_line = line.strip()
-                        break
-                if err_line:
-                    print_info(f"GHCP recommendation command failed: {err_line}")
-                else:
-                    print_info(f"GHCP recommendation command failed (exit {res.returncode})")
-                return []
-
-            parsed = self._parse_ghcp_recommended_profiles(res.stdout or "", allowed_profiles)
-            if not parsed:
-                print_info("GHCP returned output, but no valid available profile names were parsed")
-            return parsed
-        except Exception:
-            print_info("GHCP recommendation call raised an exception")
-            return []
-
-    def _parse_ghcp_recommended_profiles(self, output_text: str, allowed_profiles: List[str]) -> List[str]:
-        """Parse GHCP response into an ordered, allowed profile list for phase 4."""
-        if not output_text:
-            return []
-
-        allowed_lower = {p.lower(): p for p in allowed_profiles}
-
-        # Preferred machine-readable format.
-        for line in output_text.splitlines():
-            if line.strip().upper().startswith("RECOMMENDED_PROFILES:"):
-                rhs = line.split(":", 1)[1]
-                tokens = [t.strip().lower() for t in rhs.split(",") if t.strip()]
-                picks = []
-                for token in tokens:
-                    profile = allowed_lower.get(token)
-                    if profile and profile not in picks:
-                        picks.append(profile)
-                if picks:
-                    return picks
-
-        # JSON fallback: {"recommended_profiles": ["node", ...]}
-        try:
-            obj = json.loads(output_text)
-            arr = obj.get("recommended_profiles") if isinstance(obj, dict) else None
-            if isinstance(arr, list):
-                picks = []
-                for item in arr:
-                    if not isinstance(item, str):
-                        continue
-                    profile = allowed_lower.get(item.strip().lower())
-                    if profile and profile not in picks:
-                        picks.append(profile)
-                if picks:
-                    return picks
-        except Exception:
-            pass
-
-        # Best-effort fallback from free text.
-        text = output_text.lower()
-        picks = []
-        for profile in allowed_profiles:
-            if re.search(r"\b" + re.escape(profile.lower()) + r"\b", text) and profile not in picks:
-                picks.append(profile)
-        return picks
-
-    def _select_profiles_for_phase4(self, hot_events: Path):
-        """Choose profile set to apply based on phase 3 diagnostics artifacts."""
-        available_profiles = self._get_available_profiles()
-        print_info(f"Available profiles from CLI: {', '.join(available_profiles)}")
-
-        python_recs = self._python_profile_recommendations(hot_events, available_profiles)
-        ghcp_recs = self._ghcp_profile_recommendations(hot_events, available_profiles)
-
-        if python_recs:
-            print_info(f"Python recommendations: {', '.join(python_recs)}")
-        else:
-            print_info("Python recommendations unavailable")
-
-        if ghcp_recs:
-            print_info(f"GHCP recommendations:   {', '.join(ghcp_recs)}")
-
-        # Default strategy: try GHCP first, then fall back to Python, then default set.
-        if ghcp_recs:
-            selected = ghcp_recs
-            source = "ghcp"
-        elif python_recs:
-            selected = python_recs
-            source = "python"
-            print_info("GHCP recommendations unavailable; falling back to Python recommendations")
-        else:
-            selected = list(available_profiles)
-            source = "default"
-            print_info("No recommendations available; using default profile set")
-
-        self.recommended_profiles = selected
-        self.recommendation_source = source
-        print_info(f"Phase 4 will apply ({source}): {', '.join(self.recommended_profiles)}")
-
     def _clean_build(self):
         """Clean build artifacts to keep before/after runs comparable."""
         cache_dir = self.config.repo_path / "node_modules" / ".cache"
@@ -405,132 +178,6 @@ class VSCodeScenario(ProfiledBuildScenario):
                 tsinfo.unlink()
             except Exception:
                 pass
-
-    def _start_cpu_monitor(self, log_file: Path):
-        """Start background CPU sampling for wdavdaemon_unprivileged."""
-        stop_event = threading.Event()
-
-        def _run():
-            with log_file.open("w") as handle:
-                while not stop_event.is_set():
-                    try:
-                        result = subprocess.run(
-                            ["ps", "-eo", "pid,%cpu,comm"],
-                            capture_output=True,
-                            text=True,
-                            timeout=3,
-                            check=False,
-                        )
-                        for line in result.stdout.splitlines():
-                            if "wdavdaemon_unprivileged" in line:
-                                parts = line.split()
-                                if len(parts) >= 2:
-                                    handle.write(f"{int(time.time())} {parts[1]}\n")
-                                    handle.flush()
-                                break
-                    except Exception:
-                        pass
-                    stop_event.wait(2)
-
-        thread = threading.Thread(target=_run, daemon=True)
-        thread.start()
-        return stop_event, thread
-
-    def _calc_avg_cpu(self, log_file: Path):
-        """Compute average sampled CPU from monitor log."""
-        if not log_file.exists() or log_file.stat().st_size == 0:
-            return "N/A"
-        vals = []
-        for line in log_file.read_text().splitlines():
-            try:
-                vals.append(float(line.split()[1]))
-            except Exception:
-                continue
-        if not vals:
-            return "N/A"
-        return f"{sum(vals)/len(vals):.1f}"
-
-    def _collect_rtp_stats(self, out_file: Path):
-        """Capture RTP statistics JSON snapshot."""
-        try:
-            with out_file.open("w") as handle:
-                subprocess.run(
-                    [
-                        "sudo",
-                        "mdatp",
-                        "diagnostic",
-                        "real-time-protection-statistics",
-                        "--output",
-                        "json",
-                    ],
-                    stdout=handle,
-                    stderr=subprocess.DEVNULL,
-                    timeout=30,
-                    check=False,
-                )
-        except Exception:
-            pass
-
-    def _count_rtp_scans(self, json_file: Path):
-        """Return total scanned files from RTP stats JSON."""
-        if not json_file.exists() or json_file.stat().st_size == 0:
-            return "N/A"
-        try:
-            data = json.loads(json_file.read_text())
-            total = 0
-            for item in data:
-                if isinstance(item, dict):
-                    total += int(item.get("totalFilesScanned", 0) or 0)
-            return str(total)
-        except Exception:
-            return "N/A"
-
-    def _start_hot_event_collection(self):
-        """Start hot-event collection to run concurrently with an active build."""
-        if os.environ.get("PYTEST_CURRENT_TEST"):
-            return None, set()
-        try:
-            existing = {str(p.resolve()) for p in self.config.repo_path.glob("hot_event_source_*.json")}
-            print_info(f"Collecting hot event sources during compile (~{self.hot_event_duration}s)...")
-            proc = subprocess.Popen(
-                [
-                    "sudo",
-                    "mdatp",
-                    "diagnostic",
-                    "hot-event-sources",
-                    f"--time={self.hot_event_duration}",
-                ],
-                cwd=self.config.repo_path,
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-            return proc, existing
-        except Exception:
-            return None, set()
-
-    def _finalize_hot_event_collection(self, proc, existing_paths, out_file: Path):
-        """Finalize hot-event capture and persist the newest produced artifact."""
-        if proc is None:
-            return False
-        try:
-            proc.wait(timeout=self.hot_event_duration + 20)
-        except Exception:
-            try:
-                proc.terminate()
-            except Exception:
-                pass
-
-        candidates = sorted(self.config.repo_path.glob("hot_event_source_*.json"), key=lambda p: p.stat().st_mtime)
-        if not candidates:
-            return False
-
-        new_candidates = [p for p in candidates if str(p.resolve()) not in existing_paths]
-        picked = new_candidates[-1] if new_candidates else candidates[-1]
-        try:
-            shutil.copy2(picked, out_file)
-            return True
-        except Exception:
-            return False
 
     def _collect_hot_event_sources(self, out_file: Path):
         """Collect hot event sources during a build and persist latest report."""
